Integrate contrib types into unified Bicep extension
Wire up the build pipeline so that contrib resource type manifests
(maintained via 'make update-resource-types') are included in the unified
'radius' Bicep extension alongside the existing TypeSpec/Swagger-generated
types. This eliminates the separate radiusCompute, radiusData, and
radiusSecurity extensions.

- Add generate-bicep-types-contrib Makefile target that reads
  deploy/manifest/defaults.yaml to discover which contrib namespaces to
  include and passes all per-type manifests for each namespace to
  'manifest-to-bicep generate' for merging.

- Add generate-yq-installed prerequisite target for yq detection.

- Add publish-bicep-extension Makefile target wrapping
  'bicep publish-extension' for local testability.

- Extract buildTypeIndex() from generate.ts into a shared
  index-builder.ts module. Add rebuild-index.ts as a thin CLI wrapper.

- Extend generate-bicep-types target to call generate-bicep-types-contrib
  then rebuild-index after the existing autorest step.

- Remove radiusCompute, radiusData, radiusSecurity from bicepconfig.json,
  install-bicep.sh, and pkg/cli/setup since these namespaces are now
  part of the single 'radius' extension.

Part of: unified Bicep extension publishing (PR 2/4)
