+++ hack/bicep-types-radius/src/generator/src/cmd/generate.ts
@@ -20,14 +20,7 @@ import { mkdir, rm, writeFile, readFile } from "fs/promises";
 import yargs from "yargs/yargs";
 import { hideBin } from "yargs/helpers";
 import { Dictionary } from "lodash";
-import {
-  TypeFile,
-  buildIndex,
-  readTypesJson,
-  writeIndexJson,
-  writeIndexMarkdown,
-  TypeSettings,
-} from "bicep-types";
+import { buildTypeIndex } from "../index-builder";
 import { GeneratorConfig, getConfig } from "../config";
 import * as markdown from "@ts-common/commonmark-to-markdown";
 import * as yaml from "js-yaml";
@@ -312,31 +305,4 @@ async function findReadmePaths(specsPath: string) {
       .split(path.sep)
       .some((parent) => parent == "resource-manager");
   });
-}
-
-async function buildTypeIndex(
-  logger: ILogger,
-  baseDir: string,
-  version: string,
-) {
-  const typesPaths = await findRecursive(baseDir, (filePath) => {
-    return path.basename(filePath) === "types.json";
-  });
-
-  const typeFiles: TypeFile[] = [];
-  for (const typePath of typesPaths) {
-    const content = await readFile(typePath, { encoding: "utf8" });
-    typeFiles.push({
-      relativePath: path.relative(baseDir, typePath),
-      types: readTypesJson(content),
-    });
-  }
-  const indexContent = await buildIndex(
-    typeFiles,
-    (log) => logOut(logger, log),
-    { name: "Radius", version: version, isSingleton: false } as TypeSettings,
-  );
-
-  await writeFile(`${baseDir}/index.json`, writeIndexJson(indexContent));
-  await writeFile(`${baseDir}/index.md`, writeIndexMarkdown(indexContent));
 }
