+++ pkg/cli/setup/application.go
@@ -49,9 +49,6 @@ resource demo 'Applications.Core/containers@2023-10-01-preview' = {
 	bicepConfigTemplate = `{
 	"extensions": {
 		"radius": "br:biceptypes.azurecr.io/radius:%s",
-		"radiusCompute": "br:biceptypes.azurecr.io/radiuscompute:%s",
-		"radiusData": "br:biceptypes.azurecr.io/radiusdata:%s",
-		"radiusSecurity": "br:biceptypes.azurecr.io/radiussecurity:%s",
 		"aws": "br:biceptypes.azurecr.io/aws:%s"
 	}
 }`
@@ -92,5 +89,5 @@ func getVersionedBicepConfig() string {
 		tag = "latest"
 	}
 
-	return fmt.Sprintf(bicepConfigTemplate, tag, tag, tag, tag, tag)
+	return fmt.Sprintf(bicepConfigTemplate, tag, tag)
 }
